requirements.txt
altgraph==0.17.5
asgiref==3.11.0
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.11
PyJWT==2.10.1
pypdfium2==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
reportlab==4.4.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.4
whitenoise==6.11.0
